Make image argument more flexible
When requesting a VM, you shouldn't have to memorize the image ID.
It's not possible to use all or part of the image's name when
starting a request. If there are multiple results for the search
term, a list of possible images is displayed and the user is asked
to clarify which image they want to start.

Closes #7