Auto Connect Support
Once a request is ready, we can automatically connect to it via a
variety of connection methods. Depending on the VM type, some or
all of these connections may be available. If there is a choice,
the user will be prompted.

In this commit, SSH is the only method supported.