Implement Extend
`vcl extend [minutes]` extends a request given a request ID and
a number of minutes. If a request ID is not given explicitly, the
user is prompted with a list of active requests that they can extend.
The extend_time is in terms of minutes, which VCL will round up to
the nearest 15 minutes.